Added Has-nr. Replaced _⊛_▷_ with nr in one usage rule for natrec.
Examples were added that show that some linearity modalities, affine
types modalities, and linear or affine types modalities allow
non-linear (non-affine) functions to be given linear (affine) types,
and furthermore some of the modalities do not allow one to give a
linear type to a certain implementation of addition. These examples
are based on an observation due to Oskar Eriksson.

The usage rule for natrec that used to be stated using _▷_⊛_ is now
instead stated using "nr". Every natrec-star operator _▷_⊛_ can be
turned into an nr function, but the nr functions take five arguments
instead of three. Custom nr functions have been defined for the
linearity modality, the affine types modality, and the linear or
affine types modality. There is no proof that the new formulations
"work" in general, but at least they handle the problematic examples
mentioned above properly.

The new nr functions are not used in the statement of the alternative
usage rule for natrec, the one that is formulated with an inequality.
If this rule is used, then problematic examples can be constructed.