Replaced _∧_ with _+_ in the usage rules for J and K.
Now the inequality ω · (p + q) ≤ ω · q must hold for every semiring
with meet.

I had thought about this change before, but had hesitated, partly
because ω · (p + q) is equal to ω · (p ∧ q) for certain key
modalities, and partly because _+_ is not necessarily decreasing.
Recently Oskar Eriksson suggested that this change should be made, and
when I looked at the problem again I realised that the condition above
suffices to enable the change. Furthermore the condition holds for all
modalities for which ω · (p + q) is equal to ω · (p ∧ q).

Note that for modalities with 𝟙 ≰ 𝟘 the value of ω is now necessarily
distinct from 𝟙. The definition of ω was changed for
Graded.Modality.Instances.Zero-below-one.𝟘≤𝟙-semiring-with-meet.