Included everything in Everything.agda, made everything type-check.
I also documented the inclusion of (opaque) definitions in the README
files.

Now definitions are supported, to varying degree, by most of the
development. The target language for extraction now supports
recursive, top-level definitions. Soundness of erasure is proved for
definition contexts in which all definitions are transparent, and
there is a counterexample for the case in which definitions do not
have to be transparent.

One could perhaps associate a mode to every top-level definition and
disallow the use of definitions with mode zero when the mode is one.
If erased matches are disallowed, then soundness for erasure could
perhaps be proved for definition contexts where every opaque
definition has mode zero. It might also make sense to add top-level
postulates, required to have mode zero.

Names are disallowed in parts of the heap development, and inlining is
used to obtain some results that work for terms that contain names.

Perhaps some things could be simplified if an extra index were added
to the type of terms: the number of allowed names. For instance, I
proved separately that several forms of reduction preserve "the state
does not contain any names". However, I do not want to make such a
change in this commit.