Remove the level annotation from the lift constructor
This allows the proof of full reduction to go through. Without this
change, the issue was that the normal form of any term of type Lift l A
should be of the form lift k t because of η-equality, where k is a
normal form equal to l; however, nothing in the Lift-η algorithmic conversion
rule mentions l.

I tried to fix this in a couple by adding an assumption that Γ ⊢ l [conv↑] l ∷ Level to Lift-η.
For this to be stable under conversion (Definition.Conversion.Conversion), this actually has to be
∃ λ k → Γ ⊢ k [conv↑] k ∷ Level × Γ ⊢ l ≡ k ∷ Level.
The issue is then in Definition.Conversion.Lift, where we must
convert Γ ⊢ t ~ u ↓ Lift l A into [conv↓], but we again have no
information about l. I tried adding an assumption that A [conv↑] A,
but this must then be threaded around and fails in the case of Σ-η.

Presumably there is a better way to structure/think about these things
so that the solution is evident, but I do not know what a "generic
equality relation" is even supposed to be. For example, should
Γ ⊢ t [conv↑] u ∷ A imply Γ ⊢ A [conv↑] A a priori?

This change also makes lift checked instead of inferred, since we cannot
infer the level any more.