Removed Neutralˡ, modified Neutral and added Neutralᵃ.
Now Neutral includes all neutral terms. A neutral is atomic if it is
not an application of _supᵘ_.

NeutralAt, NfNeutral and _⊢ne_∷_ were also modified, and NfNeutralˡ
and _⊢neˡ_∷_ were removed.

The predicates Functionᵃ, Productᵃ and Identityᵃ were added, and
Identity-rec was replaced by Identityᵃ-rec.

The lemmas Neutral→Neutralᵃ-⊢∷ and Neutral→Neutralᵃ-⊢ were added.

Some predicates (like "Type" and "Function") that previously were
satisfied only for those neutral terms that are atomic neutral are now
satisfied for all neutral terms. Similarly, some lemmas that
previously took an argument or produced a value of type "Neutral t"
still do that, but now that type is inhabited for more terms t.