Now constraints are checked immediately instead of being collected.
Agda had trouble handling an example with 4483 constraints, even
though only three were distinct. I think it would be fair to say that
the code had a space leak.

The old code had the advantage that the type-checker informed the user
of exactly which constraints were required. Now one can get rather
uninformative error messages, but it should not be too hard to improve
them if that turns out to be useful.